+++ packages/xchain-bitcoin/__tests__/utils.test.ts
@@ -74,7 +74,10 @@ describe('Bitcoin Utils Test', () => {
 
   describe('broadcastTx', () => {
     it('returns txHash', async () => {
-      const txHash = await Utils.broadcastTx({ txHex: '0xdead', haskoinUrl: 'haskoin-url' })
+      const txHash = await Utils.broadcastTx({
+        txHex: '0xdead',
+        haskoinUrl: 'haskoin-url',
+      })
       expect(txHash).toEqual('mock-txid')
     })
   })