+++ packages/xchain-litecoin/__mocks__/thornode-api.ts
@@ -22,13 +22,16 @@ export default {
     })
 
     // Mock ltc node send tx
-    mock.onPost(/ltc.thorchain.info/).reply(() => [
-      200,
-      {
-        id: '1',
-        result: 'mock-txid',
-        error: null,
-      },
-    ])
+    mock.onPost(/ltc.thorchain.info/).reply((req) => {
+      console.log(req.headers)
+      return [
+        200,
+        {
+          id: '1',
+          result: 'mock-txid',
+          error: null,
+        },
+      ]
+    })
   },
 }