+++ packages/xchain-thorchain-query/__e2e__/thornode-checkTx-e2e.ts
@@ -1,6 +1,25 @@
-import { ThorchainCache } from '../src/thorchain-cache'
-import { TransactionStage, TxType } from '../src/thorchain-checktx'
+// import { Network } from '@xchainjs/xchain-client'
+// import axios from 'axios'
 
+import { ThorchainCache } from '../src/thorchain-cache'
+<<<<<<< HEAD
+import { TransactionStage, TxType } from '../src/thorchain-checktx'
+=======
+import { TransactionStage } from '../src/thorchain-checktx'
+// import { MidgardConfig } from '../src/types'
+// import { Midgard } from '../src/utils/midgard'
+// import { Thornode, ThornodeConfig } from '../src/utils/thornode'
+>>>>>>> master
+
+// const mc: MidgardConfig = {
+//   apiRetries: 2,
+//   midgardBaseUrls: ['https://eoiulqma2feqa8x.m.pipedream.net'],
+// }
+// const tc: ThornodeConfig = {
+//   apiRetries: 2,
+//   thornodeBaseUrls: ['https://eoiulqma2feqa8x.m.pipedream.net'],
+// }
+// const thorchainCache = new ThorchainCache(new Midgard(Network.Mainnet, mc), new Thornode(Network.Mainnet, tc))
 const thorchainCache = new ThorchainCache()
 
 const checkTxStage = new TransactionStage(thorchainCache)
@@ -10,6 +29,16 @@ const checkTxStage = new TransactionStage(thorchainCache)
 //   //  '991DFE33AC4482CC7A1E3BF1142E121A315EED18ED8E8FDDDC678E8F176DFCBA', // THOR in, ETH out
 //   '619F2005282F3EB501636546A8A3C3375495B0E9F04130D8945A6AF2158966BC', // BTC in, Synth BTC out
 // ]
+
+// axios.interceptors.request.use((request) => {
+//   console.log('Starting Request', JSON.stringify(request, null, 2))
+//   return request
+// })
+
+// axios.interceptors.response.use((response) => {
+//   console.log('Response:', JSON.stringify(response, null, 2))
+//   return response
+// })
 
 describe('Thorchain query checkTx Integration Tests', () => {
   it(`Should check asymBTCAddLp `, async () => {